add flask api predict NER
