+++ data_loader.py
@@ -1,0 +1,158 @@
+import copy
+import json
+
+
+class InputExample(object):
+    """
+    A single training/test example for simple sequence classification.
+
+    Args:
+        guid: Unique id for the example.
+        words: list. The words of the sequence.
+        intent_label: (Optional) string. The intent label of the example.
+        slot_labels: (Optional) list. The slot labels of the example.
+    """
+
+    def __init__(self, guid, words, slot_labels=None):
+        self.guid = guid
+        self.words = words
+        self.slot_labels = slot_labels
+
+    def __repr__(self):
+        return str(self.to_json_string())
+
+    def to_dict(self):
+        """Serializes this instance to a Python dictionary."""
+        output = copy.deepcopy(self.__dict__)
+        return output
+
+    def to_json_string(self):
+        """Serializes this instance to a JSON string."""
+        return json.dumps(self.to_dict(), indent=2, sort_keys=True) + "\n"
+
+
+class InputFeatures(object):
+    """A single set of features of data."""
+
+    def __init__(self, input_ids, attention_mask, token_type_ids, slot_labels_ids):
+        self.input_ids = input_ids
+        self.attention_mask = attention_mask
+        self.token_type_ids = token_type_ids
+        self.slot_labels_ids = slot_labels_ids
+
+    def __repr__(self):
+        return str(self.to_json_string())
+
+    def to_dict(self):
+        """Serializes this instance to a Python dictionary."""
+        output = copy.deepcopy(self.__dict__)
+        return output
+
+    def to_json_string(self):
+        """Serializes this instance to a JSON string."""
+        return json.dumps(self.to_dict(), indent=2, sort_keys=True) + "\n"
+
+
+def convert_examples_to_features(
+        examples,
+        max_seq_len,
+        tokenizer,
+        pad_token_label_id=-100,
+        cls_token_segment_id=0,
+        pad_token_segment_id=0,
+        sequence_a_segment_id=0,
+        mask_padding_with_zero=True,
+):
+    # Setting based on the current model type
+    cls_token = tokenizer.cls_token
+    sep_token = tokenizer.sep_token
+    unk_token = tokenizer.unk_token
+    pad_token_id = tokenizer.pad_token_id
+
+    features = []
+    for (ex_index, example) in enumerate(examples):
+
+        # Tokenize word by word (for NER)
+        tokens = []
+        slot_labels_ids = []
+        for word, slot_label in zip(example.words, example.slot_labels):
+            word_tokens = tokenizer.tokenize(word)
+            if not word_tokens:
+                word_tokens = [unk_token]  # For handling the bad-encoded word
+            tokens.extend(word_tokens)
+            # Use the real label id for the first token of the word, and padding ids for the remaining tokens
+            slot_labels_ids.extend([int(slot_label)] + [pad_token_label_id] * (len(word_tokens) - 1))
+
+        # Account for [CLS] and [SEP]
+        special_tokens_count = 2
+        if len(tokens) > max_seq_len - special_tokens_count:
+            tokens = tokens[: (max_seq_len - special_tokens_count)]
+            slot_labels_ids = slot_labels_ids[: (max_seq_len - special_tokens_count)]
+
+        # Add [SEP] token
+        tokens += [sep_token]
+        slot_labels_ids += [pad_token_label_id]
+        token_type_ids = [sequence_a_segment_id] * len(tokens)
+
+        # Add [CLS] token
+        tokens = [cls_token] + tokens
+        slot_labels_ids = [pad_token_label_id] + slot_labels_ids
+        token_type_ids = [cls_token_segment_id] + token_type_ids
+
+        input_ids = tokenizer.convert_tokens_to_ids(tokens)
+
+        # The mask has 1 for real tokens and 0 for padding tokens. Only real
+        # tokens are attended to.
+        attention_mask = [1 if mask_padding_with_zero else 0] * len(input_ids)
+
+        # Zero-pad up to the sequence length.
+        padding_length = max_seq_len - len(input_ids)
+        input_ids = input_ids + ([pad_token_id] * padding_length)
+        attention_mask = attention_mask + ([0 if mask_padding_with_zero else 1] * padding_length)
+        token_type_ids = token_type_ids + ([pad_token_segment_id] * padding_length)
+        slot_labels_ids = slot_labels_ids + ([pad_token_label_id] * padding_length)
+
+        assert len(input_ids) == max_seq_len, "Error with input length {} vs {}".format(len(input_ids), max_seq_len)
+        assert len(attention_mask) == max_seq_len, "Error with attention mask length {} vs {}".format(
+            len(attention_mask), max_seq_len
+        )
+        assert len(token_type_ids) == max_seq_len, "Error with token type length {} vs {}".format(
+            len(token_type_ids), max_seq_len
+        )
+        assert len(slot_labels_ids) == max_seq_len, "Error with slot labels length {} vs {}".format(
+            len(slot_labels_ids), max_seq_len
+        )
+
+        features.append(
+            InputFeatures(
+                input_ids=input_ids,
+                attention_mask=attention_mask,
+                token_type_ids=token_type_ids,
+                slot_labels_ids=slot_labels_ids,
+            )
+        )
+
+    return features
+
+
+def _create_examples(texts, label):
+    """Creates examples for the training and dev sets."""
+    examples = []
+    guid = "predict"
+    # 1. input_text
+    words = texts.split(" ")  # Some are spaced twice
+    slots = ["O"] * len(words)
+
+    # 2. slot
+    slot_labels = []
+    for s in slots:
+        slot_labels.append(label.index(s) if s in label else slot_labels.index("O"))
+    try:
+        assert len(words) == len(slot_labels)
+    except:
+        print(words)
+        print(slot_labels)
+        print(len(words))
+        print(len(slot_labels))
+    examples.append(InputExample(guid=guid, words=words, slot_labels=slot_labels))
+    return examples
