+++ model/vihnbert.py
@@ -1,0 +1,56 @@
+from torch import nn
+from transformers.models.roberta.modeling_roberta import RobertaPreTrainedModel, RobertaModel
+
+from model.module import SlotClassifier
+from pytorchcrf import CRF
+
+
+class ViHnBERT(RobertaPreTrainedModel):
+    def __init__(self, config, args, slot_label_lst):
+        super(ViHnBERT, self).__init__(config)
+        self.args = args
+        self.num_slot_labels = len(slot_label_lst)
+        self.roberta = RobertaModel(config)  # Load pretrained bert
+
+        self.slot_classifier = SlotClassifier(
+            config.hidden_size,
+            self.num_slot_labels,
+            args.dropout_rate,
+        )
+
+        if args.use_crf:
+            self.crf = CRF(num_tags=self.num_slot_labels, batch_first=True)
+
+    def forward(self, input_ids, attention_mask, token_type_ids=None, slot_labels_ids=None):
+        outputs = self.roberta(
+            input_ids, attention_mask=attention_mask, token_type_ids=token_type_ids
+        )  # sequence_output, pooled_output, (hidden_states), (attentions)
+        sequence_output = outputs[0]
+        pooled_output = outputs[1]  # [CLS]
+
+        slot_logits = self.slot_classifier(sequence_output)
+
+        total_loss = 0
+
+        # 2. Slot Softmax
+        if slot_labels_ids is not None:
+            if self.args.use_crf:
+                slot_loss = self.crf(slot_logits, slot_labels_ids, mask=attention_mask.byte(), reduction="mean")
+                slot_loss = -1 * slot_loss  # negative log-likelihood
+            else:
+                slot_loss_fct = nn.CrossEntropyLoss()
+                # Only keep active parts of the loss
+                if attention_mask is not None:
+                    active_loss = attention_mask.view(-1) == 1
+                    active_logits = slot_logits.view(-1, self.num_slot_labels)[active_loss]
+                    active_labels = slot_labels_ids.view(-1)[active_loss]
+                    slot_loss = slot_loss_fct(active_logits, active_labels)
+                else:
+                    slot_loss = slot_loss_fct(slot_logits.view(-1, self.num_slot_labels), slot_labels_ids.view(-1))
+            total_loss += slot_loss
+
+        outputs = ((slot_logits),) + outputs[2:]  # add hidden states and attention if they are here
+
+        outputs = (total_loss,) + outputs
+
+        return outputs
