+++ requirements.txt
@@ -1,4 +1,7 @@
 numpy==1.22.0
 pytorchcrf==1.2.0
-torch==1.9.1
-transformers==4.11.3
+torch==1.13.0
+flask==2.2.2
+flask-cors==3.0.10
+flask-restful==0.3.9
+transformers==4.25.1
